gpg: set full --dry-run when used with --show-keys
* g10/gpg.c (main): ensure that opt.dry_run matches
opt.import_options[IMPORT_DRY_RUN].
--

It seems that the import_options IMPORT_DRY_RUN bit doesn't have the
same power as the opt.dry_run bit itself.  When one is set or cleared
by command-line options or other option parsing, we should ensure that
the other is also set or cleared.

It would probably be cleaner to have just a single bit, stored in a
single place, but that kind of overhaul is beyond the scope of this
bugfix.

GnuPG-Bug-id: 4017
Signed-off-by: Daniel Kahn Gillmor <dkg@fifthhorseman.net>